Update and rename bullet_types.py to bullets.py
